% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmwm-package.R
\docType{package}
\name{wmwm-package}
\alias{wmwm}
\alias{wmwm-package}
\title{wmwm: Perform Wilcoxon-Mann-Whitney test with missing data}
\description{
Perform Wilcoxon-Mann-Whitney test in the presence of missing data with controlled Type I error regardless of the values of missing data.
}
\author{
\strong{Maintainer}: Yijin Zeng \email{yijin.zeng20@imperial.ac.uk}

Authors:
\itemize{
  \item Dean Bodenham \email{deanbodenhambsse@gmail.com}
  \item Niall Adams \email{n.adams@imperial.ac.uk}
}

}
\keyword{internal}
